+++ src/components/Bookings/ReservationModal.tsx
@@ -1,116 +1,88 @@
-import React, { useEffect, useState } from 'react';
-import '../../stylings/ReservationButtons.css'
-import '../../stylings/ModalStyling.css'
-import { collection, addDoc, getDocs, Timestamp } from "firebase/firestore";
-import { db } from "../../../firebase";
+import React, { useState } from 'react';
 import Modal from 'react-modal';
+import { Timestamp, addDoc, collection } from 'firebase/firestore';
+import { db } from '../../../firebase';
 
 interface ReservationModalProps {
     isOpen: boolean;
     onClose: () => void;
+    onBookingCreated: () => void;
 }
 
-export default function ReservationModal({onClose, isOpen}: ReservationModalProps) {
-    const ref = collection(db, "bookings");
-
+const ReservationModal: React.FC<ReservationModalProps> = ({ isOpen, onClose, onBookingCreated }) => {
     const [startTime, setStartTime] = useState('');
     const [endTime, setEndTime] = useState('');
-
-    // const vehicle_dropdown = () => {
-    //     const[vehicles, setVehicles] = useState([]);
-    //     const[selectedVehicle, setSelectedVehicle] = useState([]);
-
-    //     useEffect(() => {
-    //         const fetchVehicles = async() => {
-    //             try{
-    //                 const vehicleCollection = collection(db, "vehicles");
-    //                 const vehicleSnapshot = await getDocs(vehicleCollection);
-    //                 const vehicleList = vehicleSnapshot.docs.map((doc) => ({
-    //                     id: doc.id,
-    //                     ...doc.data()
-    //                 }));
-    //                 setVehicles(vehicleList);
-    //             } catch( error) {
-    //                 console.error("Error fetching vehicles: ", error);
-    //             }
-    //             
-    //         };
-    //        fetchVehicles();
-    //     }, []);
-    // }
+    const bookingRef = collection(db, "bookings");
 
     const handleSubmit = async (e: React.FormEvent) => {
         e.preventDefault();
-        console.log('Start Time:', startTime);
-        console.log('End Time:', endTime);
-
         if (startTime && endTime) {
             const start = new Date(startTime);
             const end = new Date(endTime);
-            let data = {
+            const bookingData = {
                 startTime: Timestamp.fromDate(start),
                 endTime: Timestamp.fromDate(end)
             };
-
             try {
-                await addDoc(ref, data);
-                console.log("Booking added Successfully!");
+                await addDoc(bookingRef, bookingData);
+                onBookingCreated();
+                onClose(); 
             } catch (error) {
-                console.error("Error adding booking: ", error);
+                console.error("Error adding booking:", error);
             }
         }
-
-        onClose();
     };
 
     return (
-        <div>
-            <Modal isOpen={isOpen} onRequestClose={onClose} style={{overlay: {zIndex:1000}}}>
-                <div className='ModalWrapper'>
-                    <div className='ModalContent'>
-                        <h2>Make a Reservation</h2>
-                        <form onSubmit={handleSubmit}>
-                            <label>
-                                Start Time:
-                                <br/>
-                                <input
-                                    type="datetime-local"
-                                    value={startTime}
-                                    onChange={(e) => setStartTime(e.target.value)}
-                                    required
-                                />
-                            </label>
-                            <br/>
-                            <br/>
-                            <label>
-                                End Time
-                                <br/>
-                                <input 
-                                    type="datetime-local" 
-                                    value={endTime} 
-                                    onChange={(e) => setEndTime(e.target.value)} 
-                                    required 
-                                />
-                            </label>
-                            <br/>
-                            {/* <label htmlFor="vehicle_dropdown">Vehicle:</label>
-                            <select id="dropdown" value={selectedOption} onChange={handleChange}>
-                                <option value="">Select an Option</option>
-                                {vehicles.map((vehicle) =>(
-                                    <option key={vehicle.id} value={vehicle.nickname}>
-                                        {vehicle.nickname}
-                                    </option>
-                                ))}
-                            </select> */}
-                            <br/>
-                            <div style={{display: 'flex',gap: '10px',justifyContent: 'center',marginTop: '20px'}}>
-                                <button type="submit">Submit</button>
-                                <button type="button" onClick={onClose}>Close</button>
-                            </div>
-                        </form>
+        <Modal
+            isOpen={isOpen}
+            onRequestClose={onClose}
+            style={{
+                overlay: { zIndex: 2000, backgroundColor: 'rgba(0, 0, 0, 0.5)' },
+                content: {
+                    position: 'fixed',
+                    top: '50%',
+                    left: '50%',
+                    transform: 'translate(-50%, -50%)',
+                    maxWidth: '500px',
+                    width: '90%',
+                    padding: '20px',
+                    borderRadius: '8px',
+                    zIndex: 2001,
+                }
+            }}
+        >
+            <div>
+                <h2>Make a Reservation</h2>
+                <form onSubmit={handleSubmit}>
+                    <label>
+                        Start Time:
+                        <input
+                            type="datetime-local"
+                            value={startTime}
+                            onChange={(e) => setStartTime(e.target.value)}
+                            required
+                        />
+                    </label>
+                    <br /><br />
+                    <label>
+                        End Time:
+                        <input
+                            type="datetime-local"
+                            value={endTime}
+                            onChange={(e) => setEndTime(e.target.value)}
+                            required
+                        />
+                    </label>
+                    <br /><br />
+                    <div style={{ display: 'flex', gap: '10px', justifyContent: 'center', marginTop: '20px' }}>
+                        <button type="submit">Submit</button>
+                        <button type="button" onClick={onClose}>Close</button>
                     </div>
-                </div>
-            </Modal>
-        </div> 
-    );   
-}
+                </form>
+            </div>
+        </Modal>
+    );
+};
+
+export default ReservationModal;
